Added InputAndChoiceWidget custom widget.
This custom widget combines a text input field with a select field. I
expect that this will be very useful, e.g. for specifying a number with
units in a form. The widget is placed in rmgweb.main.widgets so that it
can be used in multiple apps (e.g. pdep, rmg).