Added input buttons for adding and deleting species and path reactions.
This is the first step toward a web-based GUI for MEASURE. The idea is to
integrate the UI for editing the network into the existing views, and add
additional supporting views when needed.

These buttons don't do anything yet except cause the page to be reloaded.
Functionality for handling these button clicks is coming soon.