Added networkAddSpecies view for adding species to a Network.
This view provides a formset for adding basic information about one or
more species to be added to the network: label, structure, ground-state
energy. Other information about the species can be added in a (to be
implemented) future form. The form validates that the labels are new.
It does not yet add any species, nor does it check that the structure
identifiers are valid.