Added networkAddPathReaction view for adding path reactions to a Network.
This view provides a formset for adding basic information about one or
more path reactions to be added to the network: reactants, products,
ground-state energy. Other information about the reaction can be added in
a (to be implemented) future form. The reactants and products are specified
using choice fields so that we restrict the possible species to those
defined in the network. The reactions are validated in several ways,
including that they are new and that they are well-formed (e.g. the
elements are balanced). The reactions are not yet added to the network.